fix(geometry): isolate remaining DAG and recognition sympy
Run gradient coprimality and curvature/Laplace DAG arithmetic in
killable workers. Monic-normalize determinant source keys so a
nonmonic factor such as 2x does not double-count an inherited x
before the 768-guard cap.
